[TASK] Set pseudo user when simulating a user group
Simulating a user group had no user set, which resulted in some things
not being simulated (for example the fluid `role` and `authenticated`
view helpers).

As the expectation just includes a `uid` > 0, we are now setting
`uid=1` for the simulated user. However, the long term solution for
future versions should be replacing the usergroup simulation by a
simulation of a concrete user (like impersonating is done in the
backend).

Resolves: #90773
Releases: master, 9.5
Change-Id: Idf380607921184edf50ab784ffa309e864e754f2
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/63794
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: Daniel Goerz <daniel.goerz@posteo.de>
Reviewed-by: Henrik Ziegenhain <henrik@ziegenhain.me>
Reviewed-by: Benni Mack <benni@typo3.org>
Reviewed-by: Daniel Goerz <daniel.goerz@posteo.de>