[BUGFIX] Do not use hard-coded magic 1337 in admin panel
This change uses a "random" frontend user ID (currently 1337)
to a very unlikely number PHP_INT_MAX to ensure that no usual
frontend user is used to preview.

This line can be removed completely once
  https://review.typo3.org/c/Packages/TYPO3.CMS/+/64441
is fixed as well.

But 1337 is a no-go (that's why it needs to into 9.5)

Resolves: #91837
Related: #90773
Releases: master, 10.4, 9.5
Change-Id: I70eebcf1220b216b3a9929a7b31eb83e999f282a
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/65219
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>