[TASK] Make status of simulate time clearer
Once a user simulated a time with the admin panel,
it has not been easy to see the status of the field
or to understand how to disable it. To make it easier
for the user to handle a separate checkbox has been added
that allows enabling or disabling the simulate time feature.

Resolves: #94596
Releases: main, 11.5
Change-Id: I309475f19f64dd8324f2466082d979d0afdbd501
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/72883
Tested-by: core-ci <typo3@b13.com>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>