[TASK] Streamline getLanguageService function
Many classes in TYPO3 have a function called
getLanguageService(). The function definition
is not always equal and some functions have
a return type or doc comment and others do not.

This patch streamlines the function as following:

* It must always use a return type declaration
* It must not contain a doc comment header

Also, the superfluous override of getLanguageService()
in some subclasses extending AbstractFormElement has
been removed.

Resolves: #96732
Releases: main
Change-Id: Ie4e7fc19b2e84dee78496b4eec674a215a81c00c
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/73261
Tested-by: core-ci <typo3@b13.com>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Oliver Klee <typo3-coding@oliverklee.de>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Benni Mack <benni@typo3.org>