[TASK] Update @typo3.icons to v3.1
With the update to v3.x of the TYPO3 Icon set, there comes
a little change in favor of understanding what sizes of
icons to use. There is no defined "default" size anymore.

While it is still usable, the new default scales with font size.
So it does not matter if you call it "default" or "auto".

The default size for icon rendering was adapted to now use
the "medium" size. That is the same size as default was before.

Fixed sizes are now defined in order:
small:  16px
medium: 32px
large:  64px

Aside from some special treatment the CSS from the icon set
is now used. It was already reworked in the icon set to make
use of CSS variables instead of compiled fixed compiled values.

Resolves: #97809
Releases: main
Change-Id: I88ff4c4f09a08461674ae5acc6b1cc9b202787b3
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/74981
Tested-by: Oliver Hader <oliver.hader@typo3.org>
Tested-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Oliver Hader <oliver.hader@typo3.org>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>