[BUGFIX] Set proper groups in AdminPanel
EXT:frontend's Preview simulation already sets the
userGroup values properly, which allows to simulate
a usergroup.

However, this change also needs to be set in EXT:adminpanel
in order to allow usergroups.

This change adds the same change (see #96381) as in
PreviewSimulator in EXT:adminpanel.

This is required as otherwise, the Frontend user would
re-set the userGroup in FrontendUserAuthentication->createUserAspect()

Resolves: #99718
Related: #96381
Releases: main, 11.5
Change-Id: I346f14e2c7bd655efe3883c3a1543bddff8364e8
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/77605
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: Oliver Bartsch <bo@cedev.de>