[BUGFIX] Add fake TS setup in admin panel for fluid
With #100759, f:translate has been changed to use the
extbase LocalizationUtility in all scopes. This
triggers the ConfigurationManager, which fetches
TypoScript setup from the Request attribute.

But TS setup is never set in fully cached page content
scenarios and there is usually no Fluid view triggered
in this scope.

The admin panel however renders in "fully cached" as
well, which then fails when trying to fetch TS setup.

We *do* want extbase to fail in this scenario, since
TS is important for extbase configuration and it needs
to rely on TS setup being set in Frontend.

The patch thus works around the issue in admin panel
directly. The situation should resolve later again,
when work on LocalizationUtility finished.

Resolves: #102638
Related: #100759
Releases: main, 12.4
Change-Id: I5cf8af6593f37299e71b8555b79bf05c0e5e9c09
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/82396
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Stefan Bürk <stefan@buerk.tech>