[TASK] Show tree depth in EXT:indexed_search statisticAction
Previously, all pages listed in the statisticAction table were listed
at the same level.
With this patch the correct tree level is now displayed within the
table, as well as the corresponding doktype icon.

In addition, a button icon and button title have been updated and the
table column width has been optimized. The title length of the page
titles and file names displayed in the table now corresponds to the
value stored in the user settings.

Resolves: #104469
Releases: main
Change-Id: I80a12c03725fc5f088945d2b1497f05c545242dd
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/85343
Tested-by: Jasmina Ließmann <minapokhalo+typo3@gmail.com>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Jasmina Ließmann <minapokhalo+typo3@gmail.com>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>