[TASK] Better handle PSR-14 event output in EXT:adminpanel
When PSR-14 event dispatching has been added in TYPO3 v10,
the admin panel has been extended to list all dispatched
events with #90265.

This was good back in v10 with a relatively small number of
events being dispatched during a casual frontend rendering
call.

Nowadays, with many hooks being transformed into events, a
casual frontend request easily dispatches thousands of events.
Trying to render them all within the admin panel needs
hundreds of kilobytes of data to be transferred. And worse,
the resulting data is processed using symfony/var-dumper that
comes with JavaScript to create a collapsible view from it.
This client side processing triggers full CPU load for seconds,
even for tiny requests like the styleguide frontend, up to a
point that tabs crash with "He's dead Jim".

It seems we can not and should not render the full events
listing with all data anymore. With thousands of them being
dispatched, it gives no usable information anyway with the
sheer amount of data.

The patch heavily reduces the amount of rendered information:
We now only count the number of single events being dispatched
and render a simple table from it, similar to the existing
SQL view. This makes the admin panel usable and snappy again.

Devs who want to works with specific events need to debug their
data anyways and now receive a simple overview to see if they
are dispatched in their scenario.

Executed commands:
> composer rm symfony/var-dumper
> composer rm symfony/var-dumper -d typo3/sysext/adminpanel --no-update

Resolves: #104725
Related: #90265
Releases: main
Change-Id: I5ac394b02533b93c40fb72b0cdc64a01c0b637de
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/86576
Tested-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: Garvin Hicking <gh@faktor-e.de>
Reviewed-by: Garvin Hicking <gh@faktor-e.de>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: core-ci <typo3@b13.com>