[TASK] Do not call non static test methods statically
PHPUnit 11 changed some of the TestCase methods,
to not be static any more.
See: https://github.com/sebastianbergmann/phpunit/issues/5213

This change in PHPUnit is not directly reflected in
PHPCsFixer yet, but our config of the fixer, that we already use
can be changed, so that the non static method calls in our tests
are rewritten.

Resolves: #107401
Releases: 13.4, main
Change-Id: I9c8ea5e580e1ca170594691a760085ac2c9f6b7f
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/90583
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: Oliver Klee <typo3-coding@oliverklee.de>
Tested-by: Helmut Hummel <typo3@helhum.io>
Reviewed-by: Helmut Hummel <typo3@helhum.io>
Reviewed-by: Oliver Klee <typo3-coding@oliverklee.de>